Move SmartSearchSharedFixture into its own file
