+++ test/Microsoft.Health.Fhir.Shared.Tests.Integration/Features/Smart/SmartSearchTests.cs
@@ -48,7 +48,7 @@ namespace Microsoft.Health.Fhir.Tests.Integration.Features.Smart
     [Trait(Traits.Category, Categories.SmartOnFhir)]
     [FhirStorageTestsFixtureArgumentSets(DataStore.All)]
 
-    public class SmartSearchTests : IClassFixture<SmartSearchTests.SmartSearchSharedFixture>
+    public class SmartSearchTests : IClassFixture<SmartSearchSharedFixture>
     {
         private readonly SmartSearchSharedFixture _smartFixture;
         private readonly FhirStorageTestsFixture _fixture;
@@ -1033,32 +1033,6 @@ namespace Microsoft.Health.Fhir.Tests.Integration.Features.Smart
         private static string CreateSmartV2TestResourceId(string scenario)
         {
             return $"smart-v2-{scenario}-{Guid.NewGuid():N}";
-        }
-
-        private static async Task<FhirTypedElementToSearchValueConverterManager> CreateFhirTypedElementToSearchValueConverterManagerAsync()
-        {
-            var types = typeof(ITypedElementToSearchValueConverter)
-                .Assembly
-                .GetTypes()
-                .Where(x => typeof(ITypedElementToSearchValueConverter).IsAssignableFrom(x) && !x.IsAbstract && !x.IsInterface);
-
-            var referenceSearchValueParser = new ReferenceSearchValueParser(new FhirRequestContextAccessor(), new FhirServerInstanceConfiguration());
-            var codeSystemResolver = new CodeSystemResolver(ModelInfoProvider.Instance);
-            await codeSystemResolver.StartAsync(CancellationToken.None);
-
-            var fhirElementToSearchValueConverters = new List<ITypedElementToSearchValueConverter>();
-
-            foreach (Type type in types)
-            {
-                // Filter out the extension converter because it will be added to the converter dictionary in the converter manager's constructor
-                if (type.Name != nameof(FhirTypedElementToSearchValueConverterManager.ExtensionConverter))
-                {
-                    var x = (ITypedElementToSearchValueConverter)Mock.TypeWithArguments(type, referenceSearchValueParser, codeSystemResolver);
-                    fhirElementToSearchValueConverters.Add(x);
-                }
-            }
-
-            return new FhirTypedElementToSearchValueConverterManager(fhirElementToSearchValueConverters);
         }
 
         private void ConfigureFhirRequestContext(
@@ -3766,89 +3740,6 @@ namespace Microsoft.Health.Fhir.Tests.Integration.Features.Smart
             Assert.DoesNotContain(results.Results, r => r.Resource.ResourceTypeName == "Appointment");
             Assert.DoesNotContain(results.Results, r => r.Resource.ResourceTypeName == "MedicationRequest");
             Assert.DoesNotContain(results.Results, r => r.Resource.ResourceTypeName == "Practitioner");
-        }
-
-        public sealed class SmartSearchSharedFixture : IAsyncLifetime
-        {
-            private readonly DataStore _dataStore;
-            private FhirStorageTestsFixture _fixture;
-            private IScoped<IFhirDataStore> _scopedDataStore;
-            private SearchParameterDefinitionManager _searchParameterDefinitionManager;
-            private ISearchIndexer _searchIndexer;
-
-            public SmartSearchSharedFixture(DataStore dataStore)
-            {
-                _dataStore = dataStore;
-            }
-
-            public FhirStorageTestsFixture Fixture => _fixture;
-
-            public async Task InitializeAsync()
-            {
-                if (ModelInfoProvider.Instance.Version != FhirSpecification.R4 &&
-                    ModelInfoProvider.Instance.Version != FhirSpecification.R4B)
-                {
-                    return;
-                }
-
-                _fixture = new FhirStorageTestsFixture(_dataStore);
-                await _fixture.InitializeAsync();
-
-                var typedElementToSearchValueConverterManager = await CreateFhirTypedElementToSearchValueConverterManagerAsync();
-                _searchIndexer = new TypedElementSearchIndexer(
-                    _fixture.SupportedSearchParameterDefinitionManager,
-                    typedElementToSearchValueConverterManager,
-                    Substitute.For<IReferenceToElementResolver>(),
-                    ModelInfoProvider.Instance,
-                    NullLogger<TypedElementSearchIndexer>.Instance);
-                _searchParameterDefinitionManager = _fixture.SearchParameterDefinitionManager;
-                _scopedDataStore = _fixture.DataStore.CreateMockScope();
-
-                await LoadBundleAsync("SmartPatientA");
-                await LoadBundleAsync("SmartPatientB");
-                await LoadBundleAsync("SmartPatientC");
-                await LoadBundleAsync("SmartPatientD");
-                await LoadBundleAsync("SmartCommon");
-
-                await UpsertResource(Samples.GetJsonSample<Medication>("Medication"));
-                await UpsertResource(Samples.GetJsonSample<Organization>("Organization"));
-                await UpsertResource(Samples.GetJsonSample<Location>("Location-example-hq"));
-            }
-
-            public async Task DisposeAsync()
-            {
-                if (_fixture != null)
-                {
-                    await _fixture.DisposeAsync();
-                }
-            }
-
-            public async Task<UpsertOutcome> UpsertResource(Resource resource, string httpMethod = "PUT")
-            {
-                resource.Meta ??= new Meta();
-                resource.Meta.LastUpdated = DateTimeOffset.UtcNow;
-
-                ResourceElement resourceElement = resource.ToResourceElement();
-
-                var rawResource = new RawResource(resource.ToJson(), FhirResourceFormat.Json, isMetaSet: false);
-                var resourceRequest = new ResourceRequest(httpMethod);
-                var compartmentIndices = Substitute.For<CompartmentIndices>();
-                var searchIndices = _searchIndexer.Extract(resourceElement);
-                var wrapper = new ResourceWrapper(resourceElement, rawResource, resourceRequest, false, searchIndices, compartmentIndices, new List<KeyValuePair<string, string>>(), _searchParameterDefinitionManager.GetSearchParameterHashForResourceType("Patient"));
-                wrapper.SearchParameterHash = "hash";
-
-                return await _scopedDataStore.Value.UpsertAsync(new ResourceWrapperOperation(wrapper, true, true, null, false, false, bundleResourceContext: null), CancellationToken.None);
-            }
-
-            private async Task LoadBundleAsync(string sampleName)
-            {
-                var smartBundle = Samples.GetJsonSample<Bundle>(sampleName);
-
-                foreach (var entry in smartBundle.Entry)
-                {
-                    await UpsertResource(entry.Resource);
-                }
-            }
         }
     }
 }
